roboclipper_match_uploader/requirements.txt
boto3==1.35.90
google_api_python_client==2.151.0
google_auth_oauthlib==1.2.1
protobuf==5.28.3
reportlab==4.2.5
svglib==1.5.1
watchdog==6.0.0
